#  Add an optional promot workgroup memory to stack memory step.