--- a/iree/compiler/Conversion/LinalgToLLVM/LinalgTileAndVectorizePass.cpp
+++ b/iree/compiler/Conversion/LinalgToLLVM/LinalgTileAndVectorizePass.cpp
@@ -18,6 +18,7 @@
 #include "mlir/Conversion/StandardToSPIRV/StandardToSPIRV.h"
 #include "mlir/Dialect/Linalg/Transforms/CodegenStrategy.h"
 #include "mlir/Dialect/Linalg/Transforms/Hoisting.h"
+#include "mlir/IR/Matchers.h"
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
@@ -26,6 +27,13 @@
 
 namespace mlir {
 namespace iree_compiler {
+
+// TODO(ataei): Use pass options instead of global llvm flags.
+static llvm::cl::opt<bool> clEnablePromoteWorkgroupToFullTiles(
+    "iree-codegen-llvm-promote-workgroup-to-full-tiles",
+    llvm::cl::desc("Enable promoting wokgroup memory to full tiles allocated "
+                   "on the stack."),
+    llvm::cl::init(false));
 
 namespace {
 template <typename LinalgOpTy>
@@ -61,10 +69,79 @@ struct TileAndVectorizeWorkgroups
 };
 }  // namespace
 
+namespace {
+/// Pattern to promote all matmul operands to memory.
+struct PromoteMatmulSubviewsPattern
+    : public linalg::LinalgPromotionPattern<linalg::MatmulOp> {
+  PromoteMatmulSubviewsPattern(MLIRContext *context,
+                               linalg::LinalgPromotionOptions options,
+                               linalg::LinalgTransformationFilter marker,
+                               PatternBenefit benefit = 1)
+      : linalg::LinalgPromotionPattern<linalg::MatmulOp>(
+            context,
+            options.setOperandsToPromote({0, 1, 2})
+                .setUseFullTileBuffersByDefault(true),
+            marker, benefit) {}
+};
+}  // namespace
+
+namespace {
+
+// TODO(ataei): Refactor this into a common utility with LinalgToSPIRV.
+Optional<Value> allocateWorkgroupMemoryOnStack(
+    OpBuilder &b, SubViewOp subview, ArrayRef<Value> boundingSubViewSize,
+    OperationFolder *folder) {
+  // Allocate the memory into the entry block of the parent FuncOp. This better
+  // aligns with the semantics of this memory which is available at the entry of
+  // the function.
+  OpBuilder::InsertionGuard guard(b);
+  FuncOp funcOp = subview->getParentOfType<FuncOp>();
+  if (!funcOp) {
+    subview.emitError("expected op to be within std.func");
+    return llvm::None;
+  }
+  b.setInsertionPointToStart(&(*funcOp.getBody().begin()));
+  // The bounding subview size is expected to be constant. This specified the
+  // shape of the allocation.
+  SmallVector<int64_t, 2> shape = llvm::to_vector<2>(
+      llvm::map_range(boundingSubViewSize, [](Value v) -> int64_t {
+        APInt value;
+        if (matchPattern(v, m_ConstantInt(&value))) return value.getSExtValue();
+        return -1;
+      }));
+  if (llvm::any_of(shape, [](int64_t v) { return v == -1; })) return {};
+  MemRefType allocType =
+      MemRefType::get(shape, subview.getType().getElementType(), {});
+  Value buffer = b.create<AllocaOp>(subview.getLoc(), allocType);
+  return buffer;
+}
+
+LogicalResult deallocateWorkgroupMemory(OpBuilder &b, Value buffer) {
+  MemRefType bufferType = buffer.getType().dyn_cast<MemRefType>();
+  if (!bufferType) return failure();
+  return success();
+}
+
+}  // namespace
+
 void TileAndVectorizeWorkgroups::runOnFunction() {
   auto funcOp = getOperation();
   MLIRContext *context = &getContext();
   CPUKernelDispatch cpuKernelDispatch;
+
+  // Promotes workgroups subviews to a full-tile allocated on the stack.
+  if (clEnablePromoteWorkgroupToFullTiles) {
+    OwningRewritePatternList promotionPatterns;
+    promotionPatterns.insert<PromoteMatmulSubviewsPattern>(
+        context,
+        linalg::LinalgPromotionOptions().setAllocationDeallocationFns(
+            allocateWorkgroupMemoryOnStack, deallocateWorkgroupMemory),
+        linalg::LinalgTransformationFilter(
+            Identifier::get(getWorkgroupMarker(), context),
+            Identifier::get(getWorkgroupMemoryMarker(), context)));
+    ViewOp::getCanonicalizationPatterns(promotionPatterns, context);
+    (void)applyPatternsAndFoldGreedily(funcOp, std::move(promotionPatterns));
+  }
 
   // Workgroup first level of tiling.
   {
@@ -80,7 +157,10 @@ void TileAndVectorizeWorkgroups::runOnFunction() {
                   cpuKernelDispatch, builder, operation);
             }),
         linalg::LinalgTransformationFilter(
-            Identifier::get(getWorkgroupMarker(), context),
+            Identifier::get(clEnablePromoteWorkgroupToFullTiles
+                                ? getWorkgroupMemoryMarker()
+                                : getWorkgroupMarker(),
+                            context),
             Identifier::get(getWorkgroupL1TileMarker(), context)));
 
     (void)applyPatternsAndFoldGreedily(funcOp, std::move(l1patterns));
@@ -120,7 +200,8 @@ void TileAndVectorizeWorkgroups::runOnFunction() {
   // Apply vectorization patterns.
   {
     OwningRewritePatternList vectorizationPatterns;
-    linalg::insertVectorizationPatterns<linalg::ContractionOpInterface>(
+    linalg::insertVectorizationPatterns<linalg::ContractionOpInterface,
+                                        linalg::CopyOp, linalg::FillOp>(
         vectorizationPatterns, context, linalg::LinalgVectorizationOptions(),
         linalg::LinalgTransformationFilter(
             Identifier::get(getVectorizeMarker(), context)));
@@ -157,8 +238,8 @@ void TileAndVectorizeWorkgroups::runOnFunction() {
     OwningRewritePatternList vectorToLoopsPatterns;
     populateVectorToSCFConversionPatterns(vectorToLoopsPatterns, context,
                                           vectorToSCFOptions);
-    // Hosit hierarchical tiling indexing and other loop invariant transfer ops
-    // computation.
+    // Hosit hierarchical tiling indexing and other loop invariant transfer
+    // ops computation.
     linalg::hoistViewAllocOps(funcOp);
     linalg::hoistRedundantVectorTransfers(funcOp);