This improves and the performance of matmuls for linalg on tensors (1.3x slower comparing to linalg-on-buffers pass)

IR snippet: https://gist.github.com/asaadaldien/efb712e4b1cc7ab4620b3a22809f0a74

This PR:
```
------------------------------------------------------------------------------------
Benchmark                                          Time             CPU   Iterations
------------------------------------------------------------------------------------
BM_dot_384x128x128/process_time/real_time      0.529 ms        0.528 ms         1335
```
Baseline linalg-on-buffers:
```
------------------------------------------------------------------------------------
Benchmark                                          Time             CPU   Iterations
------------------------------------------------------------------------------------
BM_dot_384x128x128/process_time/real_time      0.396 ms        0.395 ms         1779

```
linalg-on-buffers + promotion:
```
------------------------------------------------------------------------------------
Benchmark                                          Time             CPU   Iterations
------------------------------------------------------------------------------------
BM_dot_384x128x128/process_time/real_time      0.473 ms        0.472 ms         1481

```