@asaadaldien 2021-02-19
> > This PR:
> > Baseline linalg-on-buffers + promotion:
> > linalg-on-buffers + promotion:
> 
> can you clarify what the second two are? I'm guessing 'this pr' is linalg-on-tensors + promotion, but what's the difference between the second two?

updated it! the last two are baseline (linalg-on-buffers) and linalg-on-buffers + promotion which seems to add ~ 0.8 ms cost.