@MaheshRavishankar 2021-02-19
> This improves and the performance of matmuls for linalg on tensors (1.3x slower comparing to linalg-on-buffers pass)
> 
> IR snippet: [gist.github.com/asaadaldien/efb712e4b1cc7ab4620b3a22809f0a74](https://gist.github.com/asaadaldien/efb712e4b1cc7ab4620b3a22809f0a74)

Just noticed in the IR an `alloca (..) : memref<?xi8>` . Could we make this an alloca of proper shape and element type?